trap, deleted_files
change ErrorType to Trap and change how checking if files are deleted. Will now just run the FileStat function on a remote file: if it returns Err(_), we can assume the file is removed from the source path. Will maybe need to change this method later as it is not practical to have it return Err(_)